@@ -15,24 +15,24 @@
 		"@typescript-eslint/parser": "^5.18.0",
 		"@uiw/react-markdown-editor": "^4.0.2",
 		"axios": "^0.25.0",
-		"react": "^18.0.0",
-		"react-dom": "*",
-		"react-hook-form": "^7.29.0",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
+		"react-hook-form": "^7.27.1",
+		"react-router-dom": "^6.3.0",
 		"react-redux": "^7.2.6",
 		"react-scripts": "5.0.0",
+		"@types/react": "^17.0.39",
+		"@types/react-dom": "^18.0.6",
+		"@types/redux-saga": "^0.10.5",
 		"redux": "^4.1.2",
 		"redux-saga": "^1.1.3",
 		"typescript": "^4.6.3",
 		"web-vitals": "^2.1.0",
 		"universal-cookie": "^4.0.4",
-		"@types/react-router-dom": "^5.3.3",
-		"react-router-dom": "^6.3.0"
+		"@types/react-router-dom": "^5.3.3"
 	},
 	"devDependencies": {
-		"@types/node": "^17.0.14",
-		"@types/react": "^17.0.39",
-		"@types/react-dom": "^17.0.11",
-		"@types/redux-saga": "^0.10.5"
+		"@types/node": "^17.0.14"
 	},
 	"scripts": {
 		"start": "react-scripts start",
